feat: relative coordinates
splits bbox into bbox and canvas. bbox and canvas have the same width
and height. However the position of the canvas is more akin to local
coordinates while the position of the bbox is the position of the child
canvas in its parent's canvas. This is super useful for e.g. specifying
axes and placing them relative to other objects. This helps
compositionality because glyphs can use local coordinates and be
oblivious to their parent.